% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_matthew.R
\name{replace_matthew}
\alias{replace_matthew}
\title{Replace all instances of Matthew with Matt}
\usage{
replace_matthew(string)
}
\arguments{
\item{string}{A string in which to convert Matthews to Matts}
}
\value{
A character vector
}
\description{
Replace all instances of Matthew with Matt
}
\examples{
replace_matthew("This is a string with a Matthew in it")
}
